editoast: move `/health` in its own submodule
